requirements.txt
asgiref==3.5.2
Django==4.1
django-crispy-forms==1.14.0 
django-widget-tweaks==1.4.12
djangorestframework==3.14.0 
psycopg2==2.9.3
python-decouple==3.6        
pytz==2022.6
sqlparse==0.4.2
tzdata==2022.2